Re-doing From the Begining
-homeActivity
-topLayout
-centerLayout
-bottomNav
