+++ ProjectCode/app/src/main/java/com/example/sakibmahmud/project_code/HomeActivity.java
@@ -1,0 +1,13 @@
+package com.example.sakibmahmud.project_code;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+public class HomeActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_home);
+    }
+}
