+++ ProjectCode/app/src/main/res/layout/layout_bottom_nagivation_view.xml
@@ -1,0 +1,19 @@
+<merge xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <RelativeLayout
+        android:layout_width="match_parent"
+        android:layout_height="50dp"
+        android:layout_alignParentBottom="true">
+
+        <android.support.design.widget.BottomNavigationView
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:id="@+id/bottomNavViewBar"
+            android:background="@color/colorPrimary">
+
+        </android.support.design.widget.BottomNavigationView>
+
+    </RelativeLayout>
+
+
+</merge>
