+++ ProjectCode/app/src/main/res/layout/layout_center_viewpager.xml
@@ -1,0 +1,17 @@
+<merge xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <RelativeLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent">
+
+       <android.support.v4.view.ViewPager
+           android:layout_width="match_parent"
+           android:layout_height="match_parent"
+           android:id="@+id/container">
+
+       </android.support.v4.view.ViewPager>
+
+    </RelativeLayout>
+
+
+</merge>
