+++ ProjectCode/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Thu Mar 21 12:27:19 BDT 2019
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-4.10.1-all.zip
