--- project-code/.idea/.name
@@ -1,1 +1,0 @@
-project-code
