--- project-code/app/google-services.json
@@ -1,55 +1,0 @@
-{
-  "project_info": {
-    "project_number": "332456945481",
-    "firebase_url": "https://hungry-e797b.firebaseio.com",
-    "project_id": "hungry-e797b",
-    "storage_bucket": "hungry-e797b.appspot.com"
-  },
-  "client": [
-    {
-      "client_info": {
-        "mobilesdk_app_id": "1:332456945481:android:f024321753f33b48",
-        "android_client_info": {
-          "package_name": "com.example.sakibmahmud.cupcake"
-        }
-      },
-      "oauth_client": [
-        {
-          "client_id": "332456945481-u90bs2ilohoen1ki89l485839l1c8kot.apps.googleusercontent.com",
-          "client_type": 1,
-          "android_info": {
-            "package_name": "com.example.sakibmahmud.cupcake",
-            "certificate_hash": "4b2a68e169c1adf16be2d1cfa8780adadc55915f"
-          }
-        },
-        {
-          "client_id": "332456945481-27cr20091uo4modtijf154qbjesd64e1.apps.googleusercontent.com",
-          "client_type": 3
-        }
-      ],
-      "api_key": [
-        {
-          "current_key": "AIzaSyAU4h1rcndQNFOjZp9Gi9uV7R2I0bAuOqo"
-        }
-      ],
-      "services": {
-        "analytics_service": {
-          "status": 1
-        },
-        "appinvite_service": {
-          "status": 2,
-          "other_platform_oauth_client": [
-            {
-              "client_id": "332456945481-27cr20091uo4modtijf154qbjesd64e1.apps.googleusercontent.com",
-              "client_type": 3
-            }
-          ]
-        },
-        "ads_service": {
-          "status": 2
-        }
-      }
-    }
-  ],
-  "configuration_version": "1"
-}
