--- project-code/app/src/main/java/com/example/sakibmahmud/cupcake/MainActivity.java
@@ -1,152 +1,0 @@
-package com.example.sakibmahmud.cupcake;
-
-import android.app.ProgressDialog;
-import android.support.annotation.NonNull;
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-import android.text.TextUtils;
-import android.view.View;
-import android.widget.Button;
-import android.widget.EditText;
-import android.widget.ProgressBar;
-import android.widget.TextView;
-import android.widget.Toast;
-
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.Task;
-import com.google.firebase.auth.AuthResult;
-import com.google.firebase.auth.FirebaseAuth;
-
-public class MainActivity extends AppCompatActivity implements View.OnClickListener {
-
-    private Button buttonRegister;
-    private EditText editTextName;
-    private EditText editTextEmail;
-    private EditText editTextPassword;
-    private EditText editTextPhone;
-    private TextView textViewSignin;
-
-    private ProgressBar progressBar;
-
-    private FirebaseAuth firebaseAuth;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-
-        editTextName = findViewById(R.id.editTextName);
-        editTextEmail = findViewById(R.id.editTextEmail);
-        editTextPassword = findViewById(R.id.editTextPassword);
-        editTextPhone = findViewById(R.id.editTextPhone);
-        progressBar = findViewById(R.id.progressBar);
-        progressBar.setVisibility(View.GONE);
-
-        firebaseAuth = FirebaseAuth.getInstance();
-
-        findViewById(R.id.buttonRegister).setOnClickListener(this);
-    }
-
-    @Override
-    protected void onStart() {
-        super.onStart();
-
-        if (firebaseAuth.getCurrentUser() != null) {
-            //handle the already login user
-        }
-    }
-
-    private void registerUser() {
-        final String name = editTextName.getText().toString().trim();
-        final String email = editTextEmail.getText().toString().trim();
-        String password = editTextPassword.getText().toString().trim();
-        final String phone = editTextPhone.getText().toString().trim();
-
-        if (name.isEmpty()) {
-            editTextName.setError("Please Enter User Name");
-            editTextName.requestFocus();
-            return;
-        }
-
-        if (email.isEmpty()) {
-            editTextEmail.setError("Please Enter Email ID");
-            editTextEmail.requestFocus();
-            return;
-        }
-
-        if (!Patterns.EMAIL_ADDRESS.matcher(email).matches()) {
-            editTextEmail.setError("Email ID Already In Use.");
-            editTextEmail.requestFocus();
-            return;
-        }
-
-        if (password.isEmpty()) {
-            editTextPassword.setError("Enter Password");
-            editTextPassword.requestFocus();
-            return;
-        }
-
-        if (password.length() < 6) {
-            editTextPassword.setError("Password should be more than 6 Character");
-            editTextPassword.requestFocus();
-            return;
-        }
-
-        if (phone.isEmpty()) {
-            editTextPhone.setError("Input Phone Number");
-            editTextPhone.requestFocus();
-            return;
-        }
-
-        if (phone.length() != 11) {
-            editTextPhone.setError("Invalid Phone Number");
-            editTextPhone.requestFocus();
-            return;
-        }
-
-
-        progressBar.setVisibility(View.VISIBLE);
-        firebaseAuth.createUserWithEmailAndPassword(email, password)
-                .addOnCompleteListener(new OnCompleteListener<AuthResult>() {
-                    @Override
-                    public void onComplete(@NonNull Task<AuthResult> task) {
-
-                        if (task.isSuccessful()) {
-
-                            User user = new User(
-                                    name,
-                                    email,
-                                    phone
-                            );
-
-                            FirebaseDatabase.getInstance().getReference("Users")
-                                    .child(FirebaseAuth.getInstance().getCurrentUser().getUid())
-                                    .setValue(user).addOnCompleteListener(new OnCompleteListener<Void>() {
-                                @Override
-                                public void onComplete(@NonNull Task<Void> task) {
-                                    progressBar.setVisibility(View.GONE);
-                                    if (task.isSuccessful()) {
-                                        Toast.makeText(MainActivity.this, getString(R.string.registration_success), Toast.LENGTH_LONG).show();
-                                    } else {
-                                        //display a failure message
-                                    }
-                                }
-                            });
-
-                        } else {
-                            Toast.makeText(MainActivity.this, task.getException().getMessage(), Toast.LENGTH_LONG).show();
-                        }
-                    }
-                });
-
-    }
-
-    @Override
-    public void onClick(View v) {
-        switch (v.getId()) {
-            case R.id.buttonRegister:
-                registerUser();
-                break;
-        }
-    }
-}
