--- project-code/app/src/main/java/com/example/sakibmahmud/cupcake/User.java
@@ -1,11 +1,0 @@
-package com.example.sakibmahmud.cupcake;
-
-public class User {
-    public String name, email, phone;
-
-    public User(String name, String email, String phone) {
-        this.name = name;
-        this.email = email;
-        this.phone = phone;
-    }
-}
