--- project-code/app/src/main/res/values/styles.xml
@@ -1,14 +1,0 @@
-<resources>
-
-    <!-- Base application theme. -->
-    <style name="AppTheme" parent="Theme.AppCompat.Light.DarkActionBar">
-        <!-- Customize your theme here. -->
-        <item name="colorPrimary">@color/colorPrimary</item>
-        <item name="colorPrimaryDark">@color/colorPrimaryDark</item>
-        <item name="colorAccent">@color/colorAccent</item>
-
-        <item name = "windowNoTitle">true</item>
-        <item name = "windowActionBar">false</item>
-    </style>
-
-</resources>
